+++ src/Enums/Exceptions/BuilderExceptionsCode.php
@@ -77,4 +77,16 @@ enum BuilderExceptionsCode: int
      * the necessary FROM statement to define the data source for the query's filtering conditions.
      */
     case NoPreviousFromStatement = 1006;
+
+    /**
+     * Indicates the use of an invalid 'not equals to' operator in a query condition.
+     *
+     * This exception code is triggered when a query attempts to use a 'not equals to' comparison
+     * operator that is not recognized or supported by the SQL standard or the query builder's logic.
+     * The SQL standard supports '!=' and '<>' as valid 'not equals to' operators. Use of any other
+     * symbols or syntax for expressing 'not equals to' conditions will result in this error.
+     * This code ensures that queries are constructed with correct and standard-compliant syntax,
+     * preventing errors during query execution.
+     */
+    case InvalidNotEqualsToOperator = 1007;
 }