+++ CHANGELOG.md
@@ -2,10 +2,10 @@
 
 All notable changes to `EchoQuery` will be documented in this file.
 
-## Last update 06-04-2024
+## Last update 07-04-2024
 
 ### Added
-- Where statement equals to (=) comparison operator implemented
+- Where statement not equals to (!=, <>) comparison operator implemented
 
 ### Deprecated
 - Nothing